Relax AwaitableStateChecker and fix AwaitableLambda
1. `MuxHelper` calls `await_set_executor()` even when the underlying awaitable
has returned `true` from `await_ready()`. The `AwaitableStateChecker` does not
allow this. We discussed several other options, but for now this is the least
intrusive "fix" to stop ASSERT-failing in `linux.debug` variants.

2. We may call `AwaitableLambda::await_early_cancel()` after calling
`AwaitableLambda:: await_set_executor()`. As far as I can tell,
AwaitableStateChecker::earlyCancelReturned() does not disallow this.
Let's not ASSERT-fail on this.