Rework Nursery::start() to accept a callable and arguments.
There's currently a guideline that async functions should not accept their parameter
by references. This is fine, though, if the coroutine emitted by an async function
is immediately awaited, and only causes problems in realistic scenarios if the coroutine
is submitted into a nursery:

    Task<void> func(const std::string&);
    co_await func("hello world"); // works

    CORRAL_WITH_NURSERY(n) {
        n.start(func("hello world")); // does not work: temporary string
                                      // gets destroyed here
        co_return corral::join;
    };

This leads to an unfortunate property: when writing an async function, whether it's okay
to accept a reference depends on the context in which the function is going to be used —
particularly, whether it's going to be awaited directly or through a nursery. While
the recommendation is to err on the safe side and only accept things by value, users
might be tempted to accept arguments by references to avoid additional overhead,
and then spend some time debugging their apps when, as a result of a totally unrelated
refactoring, the coroutine gets submitted to a nursery.

We've already worked around another C++ core guideline (that async lambdas should not
have any captures), as we allow the lambda itself to be passed to `Nursery::start()`,
and its lifetime would be extended by the nursery until the coroutine completes.
We can extend this approach and have `Nursery::start()` accept a callable
and its arguments, and likewise retain a copy of the arguments until the function
completes.

`Nursery::start()` will now spawn a helper coroutine holding the callable (as it does now)
and all arguments. By making sure the invocation of the callable resides in the same full
expression as `co_await`'ing on it, we can rely on any temporaries created for the purpose
of the function call to have their lifetime extended until the coroutine completes.

Another note: as `Nursery::start()` accepts arguments by value
(`void Nursery::start(auto callable, auto... args)`, it will by default make copies
of any arguments supplied to it. It is still possible to pass a reference to it,
but it has to be explicitly requested via wrapping the reference into `std::ref()`
(or `std::cref()`):

    Task<void> func(const std::string&);

    std::string ex;
    CORRAL_WITH_NURSERY(n) {
        n.start(func, "hello world");  // works: a string is constructed implicitly
                                       // from a string literal

        std::string s("hello world");
        n.start(func, s);              // works: s is copied

        n.start(func, std::cref(ex));  // works: a string defined outside the nursery block
                                       // is passed by reference

        n.start(func, std::cref(s));   // use-after-free, don't do this
    };

Another minor but nice side benefit is that it brings `Nursery::start()` interface closer
to other well-known operations which spawn tasks, like `std::thread` constructor
or python's `trio.nursery.start_soon()`.