Allow awaitable chaining.
Corral best practices advocate for using coroutines for encapsulating
asynchronous control flow, but each invocation of a coroutine requires a frame
allocation on heap, which introduces some overhead. In certain, most trivial,
circumstances one may want to save on such an allocation.

Some most trivial use cases can be optimized by having a non-coroutine function
returning `Task<>` which may do some synchronous stuff and then delegate
to a coroutine. This technique cannot be used, however, if one needs
to transform a value returned by an awaitable, or for only slightly less
trivial use cases "run these two awaitables in this order".

For example, in this example:

   Task<int> fourty_two() {
        // ...complex asynchronous calculations...
        co_return 42;
    }
    Task<int> fourty_three() {
         int ret = co_await fourty_two();
         co_return ret + 1;
     }
-- the latter coroutine only exists to arrange a trivial synchronous operation
to be performed after complex work has been done, and it comes with its own
frame allocation.

This diff adds a new combiner, `corral::then()`, which can be used to pack
a sequence of awaitables into one awaitable. This would allow rewriting
the above example as

    Awaitable<int> auto fourty_three() {
        return fourty_two() | then([](int x) { return just(x + 1); });
    }

Note that argument function to `then()` is supposed to return an awaitable,
so `just()` (or `noop()`) is required if the lambda only does synchronous
transformations, like in the example above.

If the first operation returns a value, its lifetime is extended until
the second awaitable completes, allowing writing something like this:

    Semaphore sem;
    Awaitable<void> auto reallyDoStuff();
    Awaitable<void> auto doStuff() {
        return sem.lock() | then(&reallyDoStuff);
    }

In this example semaphore lock will be held while `reallyDoStuff` runs,
and released afterwards.

`async::noop()` and `async::just()` have been rewritten to return (trivial)
awaitables, which can be casted to `Task<>` if necessary;
so `Awaitable<int> auto giveMeInt() { return just(42); }` won't cause
any heap allocations.

Any more complex control flow primitives -- conditions and loops -- have not
been supported and are unlikely to be going to, since that quickly makes code
unreadable. Using coroutines is still recommended for such use cases,
as resulting code is typically much easier to read.