Handle destruction of UnsafeNursery from inside an async task
If the UnsafeNursery's executor is already scheduled to run, then destroying
the UnsafeNursery leaves a dangling reference to it. This currently causes
an assertion failure.

Switch to eliminating the dangling reference instead, since valid code
can break under the previous semantics. This upgrades the `scheduled_` bool
to a pointer, but the extra storage was previously padding
so there should be no impact on memory footprint.