Supporting refactoring for error policies in nurseries.
This is a supporting change for the follow-up commit, which requires a Nursery
to be a `TaskParent<void>` or a `TaskParent</*some error-conveying type*/>`
depending on the used error policy, and therefore define either `storeSuccess()`
or `storeValue()`. This diff replaces `storeSuccess()` with
`storeValue(detail::Void)` for the former case, so `TaskParent<T>` can instead
unconditionally define `storeValue(/*TMP magic*/) override`, which will work
for either case.

A matching change has been done to `Result<T>`.