Add support for error policies in nurseries.
This extends changes outlined in 9bdec5a (see there for motivation)
to `Nursery`.

Behavior of `Nursery` class is not going to change, but `Nursery` becomes
a specialization of class `BasicNursery`, which is now parametrized
by an error handling strategy.

If a non-default strategy is employed, `CORRAL_WITH_BASIC_NURSERY()` returns
a value, which can (and likely should) be examined to figure out whether
any task in the nursery failed.

Infallible tasks (i.e. `Task<void>`) can still be submitted to a nursery
with a custom error policy; such tasks are considered to always succeed
(or never complete).

Other supporting changes:

  - A bespoke `exception_ptr` value used to signal requested cancellation
    of the nursery, which is no longer available (and a similar value
    in a custom policy may not necessarily be available either); so requested
    cancellation (by an error or through `cancel()` call) is now stored
    in a lower bit of `Nursery::executor_` pointer.

  - `openNursery()` is now essentially a template, parameterized by the type
    of the pointer to nursery, and deriving the error policy from there.
    Since function templates cannot be submitted to the nursery without
    annoying typing, the function has been reworked into an object.

  - `TaskStarted<>` does not take an error policy and therefore cannot hold
    a reference to `Nursery::StartAwaiter`, so nursery-invariant portion
    of `StartAwaiter` has been reduced to `TaskStartedSink` with one virtual
    function. This results in one extra virtual function call
    per `TaskStarted` invocation.

  - `BasicNursery::StartAwaiter` has been moved out of `BasicNursey`
    to `namespace corral::detail` to address a cryptic build failure on MSVC.

  - `NurseryStartAwaiter::Maker` has been used to ensure pointer stability
    of `NurseryStartAwaiter`.