Add a degenerate thread pool.
Sometimes it may be desired to run a blocking function in a thread pool
or inline in the main thread (if using threads has been disallowed
in a config file of sorts, or if the binary/library is built without
threading support).

To address this, ThreadPool can be constructed with zero thread count,
in which case it will run tasks inline. Additionally, a new class,
DegenerateThreadPool is introduced, which is API-compatible with
ThreadPool, but always runs tasks inline, and therefore does not depend
on threading-support parts of STL, like <thread> or <atomic>.

One can therefore set up a wrapper like

   #ifdef HAVE_MULTITHREADING
   #include <corral/ThreadPool.h>
   using MyThreadPool = corral::ThreadPool;
   #else
   using MyThreadPool = corral::DegenerateThreadPool;
   #endif

and use `MyThreadPool` as a best-effort thread pool throughout the
program.

Also fix ThreadPool to correctly handle functions which return
references, and enhance test coverage.