corral::asio_awaitable: SFINAE-Friendly Initiate
The return type of the initiate member function of the specialization
of async_result for corral::asio_awaitable is specified via automatic return
type deduction. This means that in order to determine that type the compiler
must substitute into the body of the function. Failures resulting from this
substitution are not SFINAE-friendly (at least not on Clang).

Due to the fact corral::asio_awaitable is a completion token which transforms
an initiating function such that it defers initiation it must decay-copy
the arguments thereto. This is in line with the formulation Chris Kohlhoff
laid out in P1943:

> "The invocation of [the] initiation may be immediate, or it may be deferred
> (e.g. to support lazily [sic] evaluation). If initiation is deferred,
> the initiation and args objects must be decay-copied and moved as required."

However there are types which are not decay-copyable, such as
boost::asio::streambuf. Given an overloaded initiating function:

    template<typename T, typename CompletionToken>
    auto some_async_function(T&&, CompletionToken&&)
        -> decltype(async_initiate<...>(...))
    { ... }

    template<typename CompletionToken>
    auto some_async_function(boost::asio::streambuf&, CompletionToken&&)
        -> decltype(async_initiate<...>(...))
    { ... }

When a call thereto occurs the compiler must build the candidate set. To do so
it must determine the return type of each overload (were this not the case
SFINAE tricks in the return type couldn't exclude functions from a candidate
set). This requires it to determine the type returned by async_initiate which,
given corral::asio_awaitable as the completion token type, requires it
to substite into the body of async_result<...>::initiate. This substitution
results in an error which is not SFINAE-friendly and therefore leads
to a compilation failure.

Added a reproducing unit test and constrained async_result<...>::initiate
for corral::asio_awaitable such that it requires that each supplemental argument
thereto is decay-copyable.

Fixes #22.